Move related classes in core to auth package
