+++ app/src/main/java/com/Found404/paypro/AuthCallbackImpl.kt
@@ -1,7 +1,7 @@
 package com.Found404.paypro
 
 import androidx.navigation.NavController
-import com.found404.core.AuthCallback
+import com.found404.core.auth.AuthCallback
 
 class AuthCallbackImpl(private val navController: NavController) : AuthCallback {
     override fun navigateTo(destination: String) {
