--- core/src/main/java/com/found404/core/AuthModule.kt
@@ -1,4 +1,0 @@
-package com.found404.core
-
-
-interface AuthModule<LoginCredentialsType, LoginResponseType> : AuthProvider<LoginCredentialsType, LoginResponseType>, AuthDisplayable
