+++ core/src/main/java/com/found404/core/auth/AuthCallback.kt
@@ -1,4 +1,4 @@
-package com.found404.core
+package com.found404.core.auth
 
 interface AuthCallback {
     fun navigateTo(destination: String)
