+++ core/src/main/java/com/found404/core/auth/AuthConfig.kt
@@ -1,4 +1,4 @@
-package com.found404.core
+package com.found404.core.auth
 
 class AuthConfig private constructor(
     var baseUrl: String = "http://default_server_ip:8086",
