+++ core/src/main/java/com/found404/core/auth/AuthModule.kt
@@ -1,0 +1,5 @@
+package com.found404.core.auth
+
+
+interface AuthModule<LoginCredentialsType, LoginResponseType> : AuthProvider<LoginCredentialsType, LoginResponseType>,
+    AuthDisplayable
