+++ core/src/main/java/com/found404/core/auth/AuthProvider.kt
@@ -1,4 +1,4 @@
-package com.found404.core
+package com.found404.core.auth
 
 interface AuthProvider<LoginCredentialsType, LoginResponseType> {
     suspend fun loginUser(
